Ported WebRTC audio quality test from pyauto to browser_tests.
For now, just porting on Linux - Windows support will follow in a later patch.

This test loads a file using a XMLHttpRequest and WebAudio. The test sets up an audio-only WebRTC call, and the file is fed through a peerconnection by connecting the WebAudio node to a peerconnection. Through a special os-dependent configuration, the machine running the test records the output of Chrome to a file. That recorded file is trimmed for silence and then compared to the original file using PESQ. Thus, we achieve a test which ensures WebRTC plays audio somewhat correctly through a PeerConnection without distorting the sound too much.

The test will force the microphone volume on the system to 100% prior to the test, since the microphone volume is sensitive and could be modified by the user or the AGC in prior tests.

R=xians@chromium.org
BUG=225698

Review URL: https://chromiumcodereview.appspot.com/21824002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@216432 4ff67af0-8c30-449e-8e8b-ad334ec8d88c