Devtools cleanup redux.
A continuation of https://codereview.chromium.org/21697002 that cleans up the
code that landed in between the time I put that change up for review and when it
landed.

In particular:

* Eliminate using directives where they don't save code.
* Declare iterators inside the relevant loops, as per Google style guide.
* Shorten.

Eliminated {} on the two single-line-body for loops that had been added, because
the rest of this file doesn't use {} on single-line-body for loops; but since
these now have a multi-line loop declaration, I'm also happy to add braces,
whichever way the review thinks is most consistent.

BUG=none
TEST=none
R=pfeldman@chromium.org, vsevik@chromium.org

Review URL: https://codereview.chromium.org/21761007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@216434 4ff67af0-8c30-449e-8e8b-ad334ec8d88c