Add RMS value from cras to feedback.
Add the RMS value from the loopback stream in cras to the feedback
report and chrome://system. This value indicates whether cras is
actually producing audio or not at that instant (as opposed to what is
coming out from hardware).

BUG=chromium:257154, chromium:215328
TEST=Built Chrome and deployed to target device. Checked
chrome://system and verified that cras_rms entry exists and is
consistent with what is being played through cras.

Review URL: https://chromiumcodereview.appspot.com/22614002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@216542 4ff67af0-8c30-449e-8e8b-ad334ec8d88c