Make CrxInstaller creation take install prompt as a scoped_ptr
This helps to clarify that callers are passing ownership of the
ExtensionInstallPrompt that they pass in. Doing this turned up some cases
in crx_installer_browsertest.cc where we were accessing the raw pointer to
the MockInstallPrompt after handing off ownership of it, so this also 
introduces a MockPromptProxy class to save the information we need in tests
but avoids holding on to a possibly invalid pointer.

BUG=269049

TBR=phajdan.jr,benjhayden,pkasting,joaodasilva,bauerb (for mechanical changes in various files)

Review URL: https://chromiumcodereview.appspot.com/22272005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@216749 4ff67af0-8c30-449e-8e8b-ad334ec8d88c