Omnibox: Allow Bundled Omnibox Field Trial to Examine Instant Extended
Make the experiments in the bundled omnibox field trial be able to 
specify whether they should on for instant-extended-enabled users, off 
for those users but on for everyone else, or on for everyone.

I tested this interactively using a local variations server.  Plus
there's improved unit tests.

BUG=267584

Review URL: https://chromiumcodereview.appspot.com/22698002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@216848 4ff67af0-8c30-449e-8e8b-ad334ec8d88c