Resubmit 21609003: Move ExtensionError to extensions/, add error limits
Added in a fix for the memory leak.

Move ExtensionError class to extensions/browser/, since it doesn't need to be in chrome/.
Limit the number of errors stored per extension to 100.
Store errors in a map, keyed by Extension ID, since that is how we will likely be accessing them.

BUG=21734

Review URL: https://chromiumcodereview.appspot.com/22647007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@216871 4ff67af0-8c30-449e-8e8b-ad334ec8d88c