Warm-up the App List when enabled from the webstore and not shown.
On Mac, the app list is not warmed up when Chrome starts. If it is not
shown, its model will not be updated with the changes due to app
install. If shown, the app is not shown at all if it is still
installing, and it will not be highlighted if shown after install
completes.

This CL Exposes a CreateForProfile function to the AppListService, which
allows an app list to be created but not shown.

BUG=267814
TEST=From a profile that has not used the app launcher (it can have
apps), install an app from the web store. Launch the app launcher.
During install, it should show install progress. After install, it
should have the item highlighted (whether first shown before or after
install completes).
TBR=xiyuan@chromium.org

Review URL: https://chromiumcodereview.appspot.com/22730003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@216954 4ff67af0-8c30-449e-8e8b-ad334ec8d88c