Omnibox: don't assume omnibox has focus in OnAfterPossibleChange().
On Linux, right clicking in the omnibox does not give it focus, so
right-clicking and pasting can change the contents without focus in
the omnibox. Take out the DCHECK that assumes this cannot happen.

Also, change the focus_source_ to assume the source was OMNIBOX in these cases.

BUG=271462
R=pkasting@chromium.org

Review URL: https://codereview.chromium.org/22795005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@217062 4ff67af0-8c30-449e-8e8b-ad334ec8d88c