Ensure sync footer visibility is always set correctly.
There is currently a race condition on updating the sync footer
visibility on the welcome page. If the web page is created after the
initial sync footer visibility state is sent, it is never sent again.

This CL ensures that on startup, the state will always be correct.

BUG=267352

Review URL: https://chromiumcodereview.appspot.com/22375004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@217090 4ff67af0-8c30-449e-8e8b-ad334ec8d88c