Fix a copy-paste error in the extension activity log code
I had a copy-paste error in some of the recently-implemented parts of
the extension activity log.  This fixes that.

I believe the bug is harmless because update_statement isn't used again
before it goes out of scope (so it's fine that it was clobbered) and
SQLite defaults to a null value for unbound parameters (so it's fine
that insert_statement wasn't initialized).

Still, I should get the code right and this fixes it.

Review URL: https://chromiumcodereview.appspot.com/22901002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@217123 4ff67af0-8c30-449e-8e8b-ad334ec8d88c