Fix infinite loop in drive::internal::ChangeListProcessor::ApplyEntryMap.
ConvertToResourceEntry was partially modifing the result data in case of
failure, which resulted in generating a chimera Drive entry triggering the
bug. Fixed the converter to keep the output unchanged in the error case.

Along the way, revive the innocent patch reverted in r217161.

BUG=270308, 260020
R=hashimoto@chromium.org

Review URL: https://codereview.chromium.org/22863003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@217201 4ff67af0-8c30-449e-8e8b-ad334ec8d88c