Add a regression test to Files.app for restoring window geometry.
This patch adds a browser test, which tests if the newly created window has the same dimensions as the previous one. Since we don't know about the screen resolution of the testing environment, we first set the window size to 640x480 and then enlarge it by 10 pixels. Note, that this test will fail if the screen resolution is smaller than 650x490. However, such small resolutions do not happen on Chrome OS.

TEST=browser_test:*restoreGeometry*
BUG=270637

Review URL: https://chromiumcodereview.appspot.com/22371013

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@217244 4ff67af0-8c30-449e-8e8b-ad334ec8d88c