cros/ash: Remove about:transparency tweak page
This is an old feature from back when we were deciding how transparent
the window captions should be.  The transparency values is not persisted
intentionally; users cannot use this as a setting feature.

Since we have long decided how transparent we want things to be,
rip this out.

BUG=270896
TEST=compiles, on Chrome OS about:transparency shows "This web page is not available"

Review URL: https://chromiumcodereview.appspot.com/22794011

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@217402 4ff67af0-8c30-449e-8e8b-ad334ec8d88c