drive: Stop parsing unnecessary ResourceEntry for GetResourceEntryByPath
To get an entry by a FilePath, what we need to do is just tracking the parent-child relationships, no need to parse the all entries in the path.

Change ResourceMetadata::FindEntryByPathSync to output resource ID and rename it to GetResourceIdByPath
Use GetResourceEntryByPath instead of FindEntrySync.

BUG=None
TEST=unit_tests

Review URL: https://chromiumcodereview.appspot.com/23175002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@217567 4ff67af0-8c30-449e-8e8b-ad334ec8d88c