Fix creating a directory in the Save-as dialog.
This regression was caused since we were listening for commands on the dialog container only. In the save-as dialog, the new folder button is on the footer, which is not a child of the dialog container. This patch adds listening for the command on the footer.

TEST=Tested manually.
BUG=272765

Review URL: https://chromiumcodereview.appspot.com/22880017

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@218255 4ff67af0-8c30-449e-8e8b-ad334ec8d88c