Add a smoke signal to count crash dump events on Windows.
Hack Alert!

This CL attempts to count the number of times the browser process generates a crash dump. It does so by sending up a smoke signal (creating a file in a smoke-signal subdirectory of the crash-dump directory) just after generating a minidump for the crash server. On browser startup, these files are scavenged and a crash dump counter is incremented for each.

R= mad@chromium.org, cpu@chromium.org
TBR= isherman@chromium.org
BUG=
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/22493007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@218282 4ff67af0-8c30-449e-8e8b-ad334ec8d88c