NaCl: Remove "#if !defined(NACL_WIN64)" that doesn't do anything
nacl_process_host.cc is never compiled with NACL_WIN64 defined because
it's not linked into nacl64.exe, so having a conditional here is
unnecessary and somewhat misleading.

This means that we're reserving the 1GB in 64-bit builds of Chrome
that use the x86-64 sandbox, though that's not harmful.  Ideally it
should be fixed, but I'm not trying to fix it in this change.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3608
TEST=none

Review URL: https://chromiumcodereview.appspot.com/23300002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@218495 4ff67af0-8c30-449e-8e8b-ad334ec8d88c