Refactored privet http API to add context to delegate
Added pointers to operations to all privet HTTP delegates, a pointer to the
parent HTTP object to all privet operations, and a name to all privet HTTP
clients, to allow users a better grasp on the privet HTTP context.

This is mostly a precursor to notifications, where we'll need potentially
multiple info operations at the same time.

BUG=245385

Review URL: https://chromiumcodereview.appspot.com/22407014

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@218641 4ff67af0-8c30-449e-8e8b-ad334ec8d88c