Keep original destination path in CopyHostedDocument.
Currently, to upload a JSON file representing a hosted document, the operation
will be translated to server side copy operation. In the operation, the copy
will be created at root first, then move to destination directory, due to
GData WAPI's limitation. If, the root directory has a file whose name is
conflicting with the file being copied, number suffix (such as "(1)") would be
accidentally appended. So, the destination path should be kept as is, till
move process.

BUG=241814
TEST=Ran unit_tests and tested manually.

Review URL: https://chromiumcodereview.appspot.com/22888004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@218681 4ff67af0-8c30-449e-8e8b-ad334ec8d88c