Never do search term replacement when user input is in progress and we're not ignoring editing.
I'm not sure if this produces any visible change.  The only case
GetSearchTerms() behaves differently on is when there's a loading page with
unknown security style, and the most important callers of
WouldReplaceSearchURLWithSearchTerms() were explicitly checking for input in
progress anyway (indeed, removing those checks was the reason for this change).

It seems like if anyone was relying on the old behavior, they're probably doing
the wrong thing.

R=mpearson@chromium.org

Review URL: https://codereview.chromium.org/22870027

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@218873 4ff67af0-8c30-449e-8e8b-ad334ec8d88c