Correctly send loadabort event when WebUI src is specified for a <webview>.
Do not try to LoadURL when WebUI src is specified in guest. This should have
always hit NOTREACHED() in RenderViewHostImpl::AllowBindings, that would
not allow WebUI bindings for guests. However the earlier check
(process->GetActiveViewCount() > 1) used to early return if there was a regular
(non WebUI) <webview> in html and a second <webview> was loaded through
JavaScript pointing to a WebUI. This CL changes the order of these two checks
so that we first make sure we are not mistakenly letting guest's binding check
bypassed.

BUG=273954
Test=WebViewTest.Shim_TestLoadAbortIllegalChromeURL passes /w correct reasoning
now.

Review URL: https://chromiumcodereview.appspot.com/22582016

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@218904 4ff67af0-8c30-449e-8e8b-ad334ec8d88c