Remove SigninDelegateObserver.
Now that the AppListViewDelegate receives signin notifications, the
SigninDelegateObserver is unnecessary. This will also fix a DCHECK
that was being failing due to ChromeSigninDelegate's implementation
of listening to signin events.

Specifically, the DCHECK would happen after signing in from the app
launcher, signing out, and then signing in again from somewhere else
(e.g chrome:settings). ChromeSigninDelegate uses a SigninTracker
which expects the signin to be in a particular state. On the second
signin the SigninTracker hasn't been reset and does not like the
state it's in.

An alternative fix to this DCHECK is to delete the SigninTracker on
signin success/failure but since this code is now unnecessary,
we may as well remove it.

BUG=271933
TEST=No functional changes

Review URL: https://chromiumcodereview.appspot.com/22825025

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@218991 4ff67af0-8c30-449e-8e8b-ad334ec8d88c