Fix ToS loading
http://crrev.com/214674 broke loading of ToS during login by adding an
invalid prefix to JS calls, causing the C++ code to call nonexistent JS
methods. This CL fixes the calls and additionally removes a layer of
wrapper methods in JS that are no longer necessary.

The CL furthermore adds a browser test that verifies the ToS are actually
being shown correctly in the UI and the "accept" button starts a session.

BUG=270267
TEST=New browser test + manual

Review URL: https://chromiumcodereview.appspot.com/23241006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@219031 4ff67af0-8c30-449e-8e8b-ad334ec8d88c