Always enable FCM on Windows.
This also re-enables loading the blacklist in tests (and adds a content_browsertest to make sure that the configuration we expect to be testing is indeed the one we are testing).

This is a stepping stone towards getting the waterfall closer to what we actually ship and closer to Aura; I will try to force threaded compositing next provided this goes well.

This CL also cleans up compositor_util.cc which was enforcing the blacklist twice.

Instead of keeping --skip-gpu-data-loading around to avoid regressing issue 190942, introduce a better global solution which involves initializing a fake gpu_info in GpuDataManagerImplPrivate::Initialize instead of querying real gpu_info (slow) when OSMesa is the GL being used (which it is in non-pure GPU tests).

Keeps --skip-gpu-data-loading around temporarily on Mac and Linux, see http://crbug.com/277242 and try failures on patch set 17 for details.

BUG=233830, 267038, 190942, 277242
TBR=jcivelli, piman

Review URL: https://chromiumcodereview.appspot.com/22198004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@219132 4ff67af0-8c30-449e-8e8b-ad334ec8d88c