Aggregate extension activity log records in memory
When using the counting policy to reduce the extension activity log
database size, identical actions on the same day are combined together
and a count field is kept.

The initial prototype wrote each action separately to the database.
This patch will aggregate identical actions together in memory, then
perform a single database update for each unique action.

The database contents should be the same with or without this patch, but
disk I/O should be reduced.

BUG=266967

Review URL: https://chromiumcodereview.appspot.com/22657002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@219155 4ff67af0-8c30-449e-8e8b-ad334ec8d88c