file_manager: Handle "open with browser" actions in file_browser_handlers.cc
The "open with browser" actions of the file manager, such as "view-in-browser",
should be handled in file_browser_handlers.cc rather than in file_manager_util.cc,
as these are special cases of file browser handlers.

With this change, we should be able to get rid of fileBrowserPrivate.viewFiles()
as fileBrowserPrivate.executeTask() is now able to handle these actions.

Note that the original code path in file_manager_util.cc (OpenItem() and helper
functions) was mainly (or only) used for opening downloaded files.

BUG=267313
TEST=Save an HTML file. Open the file from the download shelf with "open" menu. Confirm it works as before
R=kinaba@chromium.org

Review URL: https://codereview.chromium.org/23049009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@219199 4ff67af0-8c30-449e-8e8b-ad334ec8d88c