Loosen strict manifest equality requirement during webstore installs
Right now right after an extension/app is updated in the webstore, there
can be a period of skew between the webstore and the servers that download
the crx file, which causes installs to fail because we generate a permissions
warning prompt from the newer version of the manifest but get served an
older version of the .crx file.

This changes chrome to allow differences in the manifest as long as those 
differences aren't considered to alter the effective severity of the 
permissions.

BUG=236099

Review URL: https://chromiumcodereview.appspot.com/23033005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@219322 4ff67af0-8c30-449e-8e8b-ad334ec8d88c