Always enable batched writes in the extension activity log
Previously, the --enable-extension-activity-log-testing command-line
flag would disable batched writes to the extension activity log
database.  This has poor performance; it is much better to batch up log
records in memory and write them in a single transaction to disk.

Update the defaults so batching is always on (except for in a few of the
unit tests where it is explicitly disabled).  When in testing mode, use
a smaller default timeout for flushing to disk (10 seconds vs. 2
minutes) to reduce the delay before data can be read from the database
and to limit data loss in the event of an unclean shutdown.

BUG=266967

Review URL: https://chromiumcodereview.appspot.com/23005025

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@219415 4ff67af0-8c30-449e-8e8b-ad334ec8d88c