Enable non resizable app windows to programatically maximize on Mac.
This change permits the resize, and also changes the behavior of the
zoom button to maximize windows to the usable screen area. The
default zoom functionality of NSWindow is no longer used, and the
zoom button is simply a trigger for the window to be resized.

BUG=263091
TEST=Manually test please on a machine with multiple displays, repeat on each:
https://chrome.google.com/webstore/detail/window-state-sample/hcbhfbnaaancmblfhdknlnojpafjohbi
to have a window that can be maximized by the green circle (zoom)
button on the titlebar and by the Maximize button in app. Try
combinations of Maximize, Fullscreen, Restore, etc.
Create a non resizable window
 ('create a new window',
   uncheck 'resizable',
   then in 'create' section press 'normal' button)
and explore maximize as well.

Review URL: https://chromiumcodereview.appspot.com/21651002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@219648 4ff67af0-8c30-449e-8e8b-ad334ec8d88c