[Files.app] Show the context menu on the unavailable folder shortcut.
Previously, an unavailable folder shortcut (shortcut which target folder is no-longer available) doesn't have a context menu, because it is expected to be removed soon.

But we removed the dialog to remove the shortcut at r219916, so we need the context menu for user to remove it manually. This patch adds the context menu and enables users to remove 
an unavailable folder shortcut.

BUG=270436
TEST=manual

Review URL: https://chromiumcodereview.appspot.com/23535004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@219997 4ff67af0-8c30-449e-8e8b-ad334ec8d88c