[Autofill] Simplify GetRawInfo() and SetRawInfo().
Require clients of these methods to pre-convert the field type to a storable
type if appropriate.  These two methods are really only intended to be used when
fetching data directly from, or writing data directly to, a data model, which
implies that the data should be being written to a storable type.

Also, adds DCHECKs to verify that these methods are being invoked correctly.

BUG=none
TEST=none
R=estade@chromium.org

Review URL: https://chromiumcodereview.appspot.com/22320017

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@220244 4ff67af0-8c30-449e-8e8b-ad334ec8d88c