Show 'Screenshots disabled' message instead of generic fail notification
Previously, if the user tries to take a screenshot on Chrome OS when this
is disabled by policy, they just receive a generic 'Failed to take screenshot'
notification. Instead, we now notify the user that screenshots are disabled by
their administrator if this is why it failed.

BUG=256840

Review URL: https://chromiumcodereview.appspot.com/23446005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@220363 4ff67af0-8c30-449e-8e8b-ad334ec8d88c