Use sequenced blocking pool task to create PNaCl's nexe temp files
This isn't necessary in the browser since if there are 2 requests for 
the same translation we don't care which one wins, 
(i.e. which one gets the temp file and does the translation and which one 
blocks to wait) but the unit tests do care because they try to hit particular
corner cases. This is why they were flaky. So just always sequence the temp
file creation, which will have minimal cost in practice.
Also re-enable the disabled tests.
R=jvoung@chromium.org
BUG=272492

Review URL: https://chromiumcodereview.appspot.com/23737002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@220457 4ff67af0-8c30-449e-8e8b-ad334ec8d88c