Swap main menu with app-specific menu when app window focused.
When an app window is focused, a new menu item is created for that app 
and appended to the main menu. All of Chrome's menu items are hidden. 
When a Chrome window is focused, the app's menu item is removed and 
Chrome's menu items are unhidden.

This is a second attempt at https://codereview.chromium.org/18089012/
which had to be reverted because it broke keyboard shortcuts.

This moves all packaged app menu item handling into a new AppShimMenuController.

BUG=168080, 276052
Focus the app window, the only item in the main menu bar should be that app. 
Focus a browser window, the main menu bar returns to normal.

Review URL: https://chromiumcodereview.appspot.com/22867009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@220540 4ff67af0-8c30-449e-8e8b-ad334ec8d88c