Added basic browser test for window.print() from Chrome apps.
This just tests the success of the call to window.print(). It does not
test whether the call leads to a print preview window and ultimately to
a printout. It will be extended over time to provide additional coverage
of the print functionality.

BUG=NONE

Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=220467

Review URL: https://chromiumcodereview.appspot.com/23054002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@220542 4ff67af0-8c30-449e-8e8b-ad334ec8d88c