FindBrowserWithWebContents now recognises DevTool's embedded WebContent
If the DevTools window is invoked in-tab, the task manager refused to add it to the list of displayed processes, because there was no tab owning this WebContent. This CL improves the check in FindBrowserWithWebContents to recognise the DevTool's WebContent as valid in this case. To avoid crashing in TabContentsResource::GetIcon(), a FaviconTabHelper::CreateForWebContents() call had to be added.

BUG=237841

Review URL: https://chromiumcodereview.appspot.com/23576007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@220549 4ff67af0-8c30-449e-8e8b-ad334ec8d88c