[Android] Support startup histogram.
This CL adds support for the
Startup.BrowserMessageLoopStartTimeFromMainEntry UMA histogram to
chromium_testshell. It does this by recording the start time as
early as possible on the Java side (well before the C++ library is even
loaded), and then reading this on the C++ side through a new JNI call.

BUG=266173

Review URL: https://chromiumcodereview.appspot.com/23443009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@220588 4ff67af0-8c30-449e-8e8b-ad334ec8d88c