ColorChooserWin::End should act like the dialog has closed
This is only a problem on Windows.

When the page closes itself while the color chooser dialog is open,
ColorChooserDialog::DidCloseDialog was called after the listener has been destroyed.

ColorChooserWin::End() will not actually close the color chooser dialog (because we can't) but act like it did so we can do the necessary cleanup.

BUG=279263
R=jschuh@chromium.org, pkasting@chromium.org

Review URL: https://codereview.chromium.org/23785003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@220639 4ff67af0-8c30-449e-8e8b-ad334ec8d88c