Add UI for network configuration when offline during kiosk app launch.
This is a first pass at implementing UI features that would allow the user
to configure the network during a kiosk app launch. A small button appears on
the splash screen after a short period if network connectivity is not detected.
Clicking this button brings up a signin screen for the owner of the device, and
after authenticating, the standard network error dialog appears.

BUG=242589
TEST=manually on device

Review URL: https://chromiumcodereview.appspot.com/23503014

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@220762 4ff67af0-8c30-449e-8e8b-ad334ec8d88c