Files.app: Extract URL values to a new file.
Currently both file_manager.js and file_manager_banner.js include the URL values and some URLs are doubled.
In addition, file_manager.js is considerably huge and we'd like to diet it.

This CL extracts the URL to the UrlValues enum type in a new file.

BUG=282287
TEST=manually

Review URL: https://chromiumcodereview.appspot.com/23536006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@220921 4ff67af0-8c30-449e-8e8b-ad334ec8d88c