Implement fileOperationUtil.copyTo().
This CL introduces the new utility method copyTo(), which is very similar to
FileSystem API's copyTo(), but supporting progress callback and cancellation.
The private methods used by the new method are not yet implemented, but adding
this now is useful to make sure that the background implementation actually
works well.

BUG=283713
TEST=Tested manually.

Review URL: https://chromiumcodereview.appspot.com/23653004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@220928 4ff67af0-8c30-449e-8e8b-ad334ec8d88c