Update infobar buttons when attempting to run an NPAPI plugin in ash.
Although UI review suggested to change the text to "Open Chrome on the Desktop", we already have (and are using in other situations) a string for "Relaunch Chrome on the Desktop".  I used this string instead to avoid having to get anything re-translated.  If we really need it to say Open, I will create a new string.

BUG=230127

Review URL: https://chromiumcodereview.appspot.com/23591005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@221042 4ff67af0-8c30-449e-8e8b-ad334ec8d88c