file_manager: getFileTasks() to return Drive apps iff all files are on Drive
Previously, fileBrowserPrivate.getFileTasks() returned Drive app tasks
regardless of whether files were on Drive or not, and the results were
filtered in JS side with ".driveApp" field.

It's much simpler for getFileTasks() to return Drive app tasks only if
all files are on Drive. This way, we can get rid of the filtering logic
in JS, as well as the ".driveApp" field.

BUG=284242
TEST=Install Pixlr Editor from webstore; Launch Files.app; Confirm that Pixlr Editor is an option for a JPG file on Drive, but not an option for a JPG file on the local disk.
TBR=benwells@chromium.org

Review URL: https://chromiumcodereview.appspot.com/23477036

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@221230 4ff67af0-8c30-449e-8e8b-ad334ec8d88c