Will now force the volume of all input devices in the WebRTC audio test.
The WebRTC audio quality test has been relying on a particular
configuration for the input device name, but it has turned out to be
painful to maintain. Specifically, it used to require that the default
input source was named "render.monitor" to figure out which device
to force the volume to 100% on, but now it will just force the volume
on all input devices it can find on the system.

The volume is forced to 100% since PESQ will be confused if the
recording is too low in volume (or even refuse to run if the recording
is done at 0% volume, which obviously will make for a bad comparison).

BUG=

Review URL: https://chromiumcodereview.appspot.com/23562002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@221251 4ff67af0-8c30-449e-8e8b-ad334ec8d88c