Fix flaky CountingPolicyTest.CapReturns test
The CountingPolicyTest.CapReturns and FullStreamUIPolicyTest.CapReturns tests are occasionally flaking because the IO is slow and we have a low timeout of 5s set. This raises the timeout to 8s and makes sure the flush happens before our timeout starts.

Shoutout to mvrable for figuring out what was going on.

BUG=285823
NOTRY=true
R=mpcomplete@chromium.org

Review URL: https://chromiumcodereview.appspot.com/23991002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@221516 4ff67af0-8c30-449e-8e8b-ad334ec8d88c