[Activity log] Strip incognito URLs entirely from the database
When logging an Action which refers to an incognito URL to the database
using the normal (non-testing) policy, clear the URL fields entirely.
Do not even leave an indication that an incognito URL had been present,
as was done previously.

This change was requested as part of the privacy review for the
extension watchdog.

BUG=244841

Review URL: https://chromiumcodereview.appspot.com/23432006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@221577 4ff67af0-8c30-449e-8e8b-ad334ec8d88c