Bookmark Undo service for multiple level undo/redo of bookmarks.
Added BookmarkUndoService that is owned by the profile to provide
multi-level undo and redo for bookmark changes.

This is a staged commit of a larger feature.

The code is not invoked anywhere other than the provided tests so
there should be no impact on the application at this time.
Further integration will continue in future commits.

BUG=126092

Review URL: https://chromiumcodereview.appspot.com/19287013

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@221943 4ff67af0-8c30-449e-8e8b-ad334ec8d88c