[Activity log] Rework extraction of URLs from argument lists
Redo the code for handling the tabs API and tab ID translation to make
it more generic.  Make it table driven: there is now a table of API
calls that can contain URL arguments, so it is easy to add additional
API calls that we want to have URLs extracted from.  Add a couple of
others I've come across in testing.

BUG=253368, 275701

Review URL: https://chromiumcodereview.appspot.com/23545012

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@222155 4ff67af0-8c30-449e-8e8b-ad334ec8d88c