Avoid a write to NULL in ReadFieldBinary
what happens is that we use vector_as_array(&data) which returns null if data is empty, but we don't check |size| whichcan be 0 so data.resize(size) still an empty vector.

This is currently happening for browser_tests.exe in the waterfall.

There might be a better fix but lots of tests are having exceptions so I want to stop that.

TBR=mnissler
BUG=none
TEST=logs are clean

Review URL: https://codereview.chromium.org/23961004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@222177 4ff67af0-8c30-449e-8e8b-ad334ec8d88c