InstantExtended(gtk): Hide top match if told to so.
CL 17114003 and 22410004 added a way for Mac and views to hide the top
omnibox match when it is a verbatim match we do not want to show. This
implements that for GTK too so it'll work on all desktop platforms.

BUG=252823
TEST=Manually. Run with no flags and verify that when you type 'faceb'
in the omnibox, the popup shows a search suggestion for 'faceb'. Run
again with the flags
--enable-instant-extended-api
--force-fieldtrials='InstantExtended/Group1 hide_verbatim:1/'
and verify that you do not see this suggestion. Also check that with the
flags set, hovering on the top suggestion works and that there are no
off-by-one problems when clicking on suggestions.

Review URL: https://chromiumcodereview.appspot.com/22679003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@222186 4ff67af0-8c30-449e-8e8b-ad334ec8d88c