network_list.js should reflect the correct logged in user type.
Updated InternetOptionsHandler and network_list.js to observe LoginState
updates, so that using UIAccountTweaks is not the only way to determine
whether or not the current user is owner. This is because |loadTimeData|
may not contain the current user login details if the logged in user type
updates after the UI fills in the UIAccountTweaks info.

BUG=285589
R=stevenjb@chromium.org

Review URL: https://codereview.chromium.org/23466015

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@222390 4ff67af0-8c30-449e-8e8b-ad334ec8d88c