Enable passing a file path to an app via its shim.
This allows opening a file with a packaged app by dropping the file the
app shim's dock icon or app bundle.
A base::FilePath param is added to the Launch and Focus IPC messages and
respective AppShimHandler methods. The file path ultimately is added to
the command line used by OpenApplication.

The shim is modified to set a custom NSApplicationDelegate early in
order to capture the call to application:openFile:. The file name is
saved and sent via LaunchApp in AppShimController::Init.

BUG=168080
TEST=Drag a file onto an app shim.
The app should start and open the file.
Drag a file onto the app shim's dock icon.
The app should open the file.

Review URL: https://chromiumcodereview.appspot.com/21005003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@222391 4ff67af0-8c30-449e-8e8b-ad334ec8d88c