file_manager: Remove FileWatchChangedEntry from fileBrowserPrivate
This is a reland of crrev.com/221982, which was reverted as it broke
DriveFileSystemExtensionApiTest.FileWatch. This patch contains a fix
for the test.

This type is not in use, and we don't plan to use this any time soon.
We should add it back when it's needed.

BUG=257916
TEST=Files.app works as before
TBR=mtomasz@chromium.org

Review URL: https://codereview.chromium.org/23532054

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@222451 4ff67af0-8c30-449e-8e8b-ad334ec8d88c