Add plumbing for "don't do search term replacement".  Will be used by Show URL.
The toolbar model gets a new flag, "search term replacement enabled", and the
omnibox classes set/clear/preserve it as necessary.  Note that there are not yet
any UI hooks to turn the flag on; those will be in the next (and final) patch.

This also renames some existing toolbar-related functions/arguments in the hope
of clarity.

Most of the complexity here comes from figuring out precisely when the flag
should be set and reset.  Basically, the flag is preserved across tab switches,
and is cleared when user input occurs, when the user hits enter on the URL to
reload it, or when a navigation occurs that is allowed to change the visible
text.  This gives users time to copy or edit a URL after showing it, but once
that's done, returns to showing search terms again.

BUG=248287
TEST=none
R=msw@chromium.org

Review URL: https://codereview.chromium.org/23629013

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@222887 4ff67af0-8c30-449e-8e8b-ad334ec8d88c