Refactors TestSystemTrayDelegate into DefaultSystemTrayDelegate (2)
Original patch was reverted as TestSystemTrayDelegate had an ASH_EXPORT, which caused compile error win-aura builder.

DefaultSystemTrayDelegate is the no-op implementation and
TestSystemTrayDelegate contains the test related bits.

This also fixes a crash in the windows side as
TestSystemTrayDelegate::Shutdown() quits the message loop, which is
the wrong thing to do.

BUG=159693 284358
TEST=none
R=oshima@chromium.org
TBR=oshima@chromium.org

Review URL: https://codereview.chromium.org/23792007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@223066 4ff67af0-8c30-449e-8e8b-ad334ec8d88c