Fix device policy recovery on CrOS login
If the device policy file cannot be loaded on login, the device enters
a state where it allows login process to proceed only if the user logging
in is the owner. To be able to successfully determine if the user is
owner, parallel authenticator has to wait until the certificates
are loaded by CertLoader. The problem is that the CertLoader starts
loading the certificates _after_ the user actually logs in. So if the
authenticator actually waited for the owner status to be resolved, the
login would hang.

This CL adds another state to LoginState::LoggedInState enum (SAFE_MODE)
in which CertLoader will be allowed to start loading the certificates.
When the authenticator detects the policy file corruption, it changes
login state to SAFE MODE (which triggers the certificate loading) and
waits for the DeviceSettingsService to determine whether the current user
is the owner.

Also, removed LoginState::GetLoggedInState. Replaced it's pre-existing
usages with LoginState::IsUserLoggedIn; and added
LoginState::IsInSafeMode to be used here.

BUG=285450
TEST=
1. manually remove device policy file
2. try logging in to a non-owner account -> should fail  
3. try logging in to the owner account -> should succeed
4. try logging in to a non-owner account again -> should succeed

Review URL: https://chromiumcodereview.appspot.com/23684033

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@223168 4ff67af0-8c30-449e-8e8b-ad334ec8d88c