Added a test for closing-window-while-printing in Chrome apps.
Closing the window being printed in a Chrome app while the print preview
panel is up currently crashes Chrome on OS X. The added test verifies
that this crash does not happen in that situation. It is initially
disabled on OS X and can be enabled once the bug causing the crash is
fixed.

This change list also strengthens the existing WindowDotPrintShouldWork
test by renaming it WindowDotPrintShouldBringUpPrintPreview and
verifying that calling window.dot.print() not only completes without
error but also brings up the print preview panel.

BUG=none

Review URL: https://chromiumcodereview.appspot.com/24104004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@223205 4ff67af0-8c30-449e-8e8b-ad334ec8d88c