Disable click to select all for search terms
This CL disables click to select all when the omnibox is displaying
search terms. This makes it easier for users to edit search terms.
This is the Mac version of crrev.com/22972003.

BUG=225762
TEST=Enabled Instant Extended. Did a search from the omnibox.
Clicked on the search terms and verified that text was not
selected. Navigated to a URL. Clicked on the URL. Verified
that the URL was selected.

Review URL: https://chromiumcodereview.appspot.com/23526044

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@223637 4ff67af0-8c30-449e-8e8b-ad334ec8d88c