Fix bookmark folder drop indicator
With r208637 the drop indicator in bookmark folder views would no
longer be visible.

The problem was that the bookmark button is opaque and obscuring the
indicator. Fix was to draw the indicator in a separate view.

BUG=293319
TEST=Created a bookmark folder in the bookmark bar. Added multiple
bookmarks in the folder. Verified that dragging them correctly showed
a drop indicator.

Review URL: https://chromiumcodereview.appspot.com/23819069

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@223774 4ff67af0-8c30-449e-8e8b-ad334ec8d88c