Fix chrome crash at NetworkStats.
The crash may happen when the ReadData timed out and deleted the object
but later on a packet arrived and OnReadComplete was started. So I changed
the binding for OnReadComplete to be a weakptr. 
Also if Read or write error happens, stop the tests right away. 

BUG=273917

Review URL: https://chromiumcodereview.appspot.com/23995021

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@223829 4ff67af0-8c30-449e-8e8b-ad334ec8d88c