[password generation] Upload possible account creation forms via Autofill
We consider that a form may be used for account creation if the first time
the saved credential is used it is on a different form, and the original form
has at least 3 text input fields.

This change does not enable uploading for GNOME or KWallet users.

BUG=240559

Review URL: https://chromiumcodereview.appspot.com/23580005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@223907 4ff67af0-8c30-449e-8e8b-ad334ec8d88c