Remove Mac-only chrome://flags flag for enabling the app launcher.
This makes the flags for enabling the app launcher consistent with
Windows (i.e. webstore-only) and removes some problems caused by the
disable flow offending the OSX Dock. Mac likes things being dragged to
Trash to "uninstall", which is currently the supported way to disable
the app launcher on OSX.

BUG=266325, 267831
TEST=In chrome://flags, the "Enable App Launcher OSX app bundle" should
not be shown.

Review URL: https://chromiumcodereview.appspot.com/23548038

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@224050 4ff67af0-8c30-449e-8e8b-ad334ec8d88c