Implement Bring All To Front for app windows. (Mac)
If the shim is available, this just calls ExtensionAppShimHandler::OnShimFocus for that app, otherwise this calls FocusWindowSet on the apps windows.

The menu item is now a doppelganger as it has a different action.

BUG=168080,276052
TEST=Start an app with multiple windows.
Put a third window between two of the app's windows.
Select Bring All To Front, should bring all windows of the app to the front.

Review URL: https://chromiumcodereview.appspot.com/24021010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@224067 4ff67af0-8c30-449e-8e8b-ad334ec8d88c