Use the short_name manifest property for app titles in the NTP (chrome://apps)
The "short_name" manifest property is now used for app titles in the
new tab page (chrome://apps). The tooltip displays the full app name.
This was implemented for consistency with app titles in the app
launcher.

BUG=288165
TEST=Test 1) Apps which specify a short_name which is different to
the name and 2) Apps which do not specify a short_name.

Review URL: https://chromiumcodereview.appspot.com/24172002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@224069 4ff67af0-8c30-449e-8e8b-ad334ec8d88c