[password generation] Always allow generated passwords to be shown.
Currently if you have a password saved from login and you create and save new
credentials on the site, they will not autofill on the login form because the
login credentials are considered a better match. This is not ideal in general,
but is particularly bad for generated passwords.

BUG=225579

Review URL: https://chromiumcodereview.appspot.com/23533069

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@224215 4ff67af0-8c30-449e-8e8b-ad334ec8d88c