Reland: App windows with ids should remember fullscreen state.
App windows with ids are restored at the previous location and state they had last. However, on some platforms fullscreen was not reported as the state to restore the window to (likely due to fullscreen state being added after that initial logic). This patch adds that support.

This patch was previously landed by zhchbin: https://src.chromium.org/viewvc/chrome?view=rev&revision=222226 but was rolled out due to test failure on Mac. Due to automation difficulties of fullscreen features, especially on Mac, we have been relying on manual only testing. I have reformed this patch with the automated test removed.

I have also modified NativeAppWindowGtk::IsFullscreenOrPending() to be resilient to having multiple window state changes in flight at once. This caused the window to not believe it was fullscreen when it was initially created and fullscreened before an async gtk state update was returned.


BUG=242790,287596
TEST=1. Install window-state-sample: https://chrome.google.com/webstore/detail/window-state-sample/hcbhfbnaaancmblfhdknlnojpafjohbi; 2. Run the app and create a new window with id. 3. Fullscreen the new window by clicking "Fullscreen" of "chrome.app.window" section; 4. Close the window (You can use "Alt+F4" etc.) and recreate the new window with the same id. 5. A fullscreen window should be shown.

Review URL: https://chromiumcodereview.appspot.com/23691078

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@224238 4ff67af0-8c30-449e-8e8b-ad334ec8d88c