[Activity log] Drop XMLHttpRequest.open from the argument whitelist
Do not keep all arguments for the XMLHttpRequest.open DOM call.  Usually
we will extract the URL argument into arg_url where it is accessible
even when arguments aren't kept.  Sometimes, however, the argument
contains a relative (not absolute) URL which can't be extracted.  In
this case, err on the side of privacy and do not log the URL at
all--rather than keeping the arguments around, where the URL is kept but
not sanitized or cleared with history.

Review URL: https://chromiumcodereview.appspot.com/23928006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@224282 4ff67af0-8c30-449e-8e8b-ad334ec8d88c