Do not show popup when crashing force-installed extension is restarted
CL 23427003 added code that will restart a force-installed extension if it
crashes and will show a popup if multiple crashes occur within a short
time. UI objected to the use of this popup bubble and would like a
notification based on the GlobalErrorBubble to be used instead.

This CL removes the popup that UI vetoed. A GlobalErrorBubble will be
added in a separate CL, likely resurrecting some of the code that is
removed in this CL.

BUG=175701
TEST=None

Review URL: https://chromiumcodereview.appspot.com/24242003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@224320 4ff67af0-8c30-449e-8e8b-ad334ec8d88c