<webview>: Reduce test flake.
This CL reduces <webview> test flake by not loading the test web server if the server is not necessary for the test. It also modifies the edit_commands test so that it no longer requires a test server.

This CL also re-enables WebViewInteractiveTest.PointerLock_PointerLockLostWithFocus in the hope of having removed flake as it no longer loads a test server.

BUG=293445

Review URL: https://chromiumcodereview.appspot.com/24244004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@224328 4ff67af0-8c30-449e-8e8b-ad334ec8d88c