[Retry] SyncFS: Support resolveLocalFileSystemURL on SyncFileSystem
Original review: https://codereview.chromium.org/23856002/

window.resolveLocalFileSystemURL is originally defined only for FileSystem API,
but it'd be nice if it works on SyncFileSystem in a similar way. For that this
change adds new IPC messages and interfaces.

Blink side change depends on this:
https://codereview.chromium.org/23537011/


BUG=177137,294502
TEST=manual (Get FileSystemURL for a file on SyncFS, and then resolve it)
TEST=unit_tests
TEST=run_webkit_tests.sh http/tests/filesystem/\*
TEST=run_webkit_tests.sh http/tests/inspector/filesystem/\*
TBR=tsepez@chromium.org
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/23526076

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@224542 4ff67af0-8c30-449e-8e8b-ad334ec8d88c